requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Uploads==0.2.1
Pillow==10.0.0
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0